requirements.txt
fastapi
uvicorn
jinja2
requests
qrcode
pillow
python-multipart
python-dotenv
